requirements.txt
streamlit
streamlit-option-menu
requests
st-annotated-text
